requirements.txt
python-dotenv>=1.0.0
requests>=2.31.0
pyautogui>=0.9.50
pyperclip>=1.8.0
# Image fallback (pyscreeze requires Pillow)
pyscreeze>=0.1.28
Pillow>=9.0.0
# Optional OpenCV fallback if Pillow/pyscreeze are unavailable on this Python version
opencv-python>=4.7.0.72
numpy>=1.24.0
pywinauto>=0.6.9
# OCR options
# Preferred: pytesseract (requires Tesseract engine installed separately)
pytesseract>=0.3.10
# Pure-Python OCR fallback (no system binary): easyocr (downloads models at runtime)
easyocr>=1.6.2
# easyocr dependencies
opencv-python-headless>=4.7.0.72
gspread>=5.8.0
google-auth>=2.20.0
